Disallow extra characters at start or end of string
Currently from_str allows bogus durations like "X3hY" to be parsed,
ignoring the extra characters "X" and "Y" at the start and the end
instead of raising an error.

Prevent this by using re.finditer instead of re.findall, which allows
us to check that the start() of the first match is at the start of the
string, and the end() of the last match is at the end of the string.

We also need to move the sign parsing to before the regex matching.

Add some tests for these cases, and make it so you can see which case
failed in the assertRaises testing.

In addition, move the regex compiling outside the function so it's only
done once on import (a common Python idiom to get a small performance
gain).